+++ src/lib.js
@@ -1,7 +1,7 @@
 import { mkdirSync, readdirSync, readFileSync, existsSync, writeFileSync, rmSync } from 'fs';
 import { execSync } from 'child_process';
 
-import * as utils     from '#utils/index.js'
+import * as utils     from '#utils/index.js';
 import * as templates from '#templates/index.js';
 import { LibSSL }     from '#ssl/lib-ssl.js';
 
@@ -61,7 +61,12 @@ export class Narnia {
       if (proxy.startsWith('.'))
         continue;
 
-      this.proxies[proxy] = JSON.parse(readFileSync(this.config.proxy_dir + proxy, 'utf8'));
+      // Skip invalid proxy files
+      const proxy_content = JSON.parse(readFileSync(this.config.proxy_dir + proxy, 'utf8'));
+      if (typeof proxy_content !== 'object' || !proxy_content.domain)
+        continue;
+
+      this.proxies[proxy] = proxy_content;
     }
   }
 
@@ -92,6 +97,7 @@ export class Narnia {
       state: 'disabled',
       address: `${address.protocol}://${address.host}:${address.port}`,
       certificate: false,
+      auto_renew: null,
       keepalive: Number.isInteger(options.keepalive) ? '' + options.keepalive : this.config.keepalive,
       additional: options.additional
         ? options.additional.split(',').map(d => d.trim())
@@ -122,6 +128,22 @@ export class Narnia {
       proxy.keepalive = '' + options.keepalive
     }
 
+    // Auto renew
+    if ( options['auto-renew']) {
+      const auto_renew = options['auto-renew'] + '';
+
+      const cases = {
+        'null': null,
+        'true': true,
+        'false': false,
+        'on': true,
+        'off': false
+      }
+
+      if (Object.hasOwn(cases, auto_renew))
+        proxy.auto_renew = cases[auto_renew]
+    }
+
     // Additional
     if ( options.additional ) {
       proxy.additional = options.additional.split(',').map(d => d.trim())
@@ -133,6 +155,7 @@ export class Narnia {
       proxy.additional = Array.from(new Set(proxy.additional.concat(additional)))
     }
 
+    // Set template
     if ( options.template ) {
       // Prepare template name in Title Case
       options.template = utils.titleCase(options.template);
@@ -142,6 +165,8 @@ export class Narnia {
         const available_templates = Object.keys(templates).join(', ');
         return { error: `Invalid template "${options.template.toLowerCase()}". Available templates: ${available_templates}` }
       }
+
+      proxy.template = options.template;
     }
 
     // Update proxy configuration
@@ -214,10 +239,10 @@ export class Narnia {
     let { proxy, path, error } = this.ensure(options, command)
     if ( error ) return { error };
 
-    return this.ssl_single( proxy, path );
+    return this.ssl_single( proxy, path, options );
   }
 
-  async ssl_single( proxy, path ) {
+  async ssl_single( proxy, path, options ) {
     if (!path) path = this.config.proxy_dir + proxy.domain;
 
     const staging = !!options?.staging;
@@ -235,6 +260,9 @@ export class Narnia {
 
     if ( !staging && result.success ) {
       proxy.certificate = result.certDate.getTime();
+
+      // Set auto renew to true if not explicitly defined
+      proxy.auto_renew = ( proxy.auto_renew == null ) ? true : false;
     }
 
     // Disable .well-known directory and reload nginx
@@ -258,16 +286,25 @@ export class Narnia {
 
     // Renew each one sequentially, to avoid being blocked
     // by Let's Encrypt servers
-    for ( const proxy in this.proxies ) {
+    for ( const domain in this.proxies ) {
+      let proxy = this.proxies[domain]
+
       // Skip if there is no certificate
       if (proxy.certificate == false) {
         console.log( `[Skip]: Proxy ${domain} with SSL not enabled` )
         continue;
       }
 
+      // Skip if there is no certificate
+      if (proxy.auto_renew == false) {
+        console.log( `[Skip]: Auto renew disabled for ${domain}` )
+        continue;
+      }
+
       // Ask for renew if there is a certificate
-      const response = await this.ssl_single(proxy)
+      const response = await this.ssl_single(proxy, null, options)
       this.print_response(response)
+      await utils.delay(1200);
     }
 
     return { success: 'SSL renew operation concluded' }
@@ -279,7 +316,7 @@ export class Narnia {
 
     // Iterate, ensuring the certificate expiration date is added when missing
     // for retro-compatibility
-    Object.entries(this.proxies).forEach(this.ensure_ssl_date)
+    Object.keys(this.proxies).forEach(this.ensure_ssl_date.bind(this))
 
     return { success: 'SSL check operation concluded' }
   }