+++ src/utils/index.js
@@ -1,4 +1,4 @@
-export { list_table }       from '#utils/list-utils.js'
-export { parseUrl }         from '#utils/url.js';
-export { titleCase }        from '#utils/title-case.js';
-export { isValidTimestamp } from '#utils/valid-timestamp.js';
+export { list_table }       		from '#utils/list-utils.js'
+export { parseUrl }         		from '#utils/url.js';
+export { titleCase }        		from '#utils/title-case.js';
+export { isValidTimestamp, delay } 	from '#utils/time.js';