+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 0.4.1 - 2026-02-10
 
 * Fixed `narnia install` to not run the check of installation on the `install` command itself
+* Various fixes in `narnia ssl:check` and `narnia ssl:renew`
+* Ignoring invalid proxy files
+* `narnia list` now shows the expiration date
+* Added `--auto-renew` flag to enable/disable automatic certificate renew for a proxy
 
 ## 0.4.0 - 2026-02-02